test: Make graphml comparer debuggable via environment variable
