+++ .git-blame-ignore-revs
@@ -1,0 +1,2 @@
+# refactor(tooling): Split .graphml tooling files & moved to tools/
+bae4196378d639d330d6f9a6eb88009dc2071435
